charge_bot/requirements.txt
psycopg2
python-telegram-bot
tabulate
